requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12.1,<=3.12.3
djangorestframework-simplejwt
django-cors-headers>=3.4.9,<=3.5.2
coverage
psycopg2-binary>=2.8
Pillow
pandas
boto3
django-storages
python-dotenv
